Moved qemu header to arch; added ramfs test
